Move embedded tbot out of the operator for resue by other integrations
