+++ integrations/lib/embeddedtbot/bot.go
