+++ integrations/lib/embeddedtbot/bot_test.go
